+++ azurerm/internal/services/media/media_services_account_resource.go
@@ -6,13 +6,15 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/mediaservices/mgmt/2018-07-01/media"
+	"github.com/Azure/azure-sdk-for-go/services/mediaservices/mgmt/2020-05-01/media"
 	"github.com/hashicorp/go-azure-helpers/response"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/suppress"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/media/parse"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tags"
 	azSchema "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tf/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
@@ -72,21 +74,60 @@ func resourceArmMediaServicesAccount() *schema.Resource {
 				},
 			},
 
-			// TODO: support Tags when this bug is fixed:
-			// https://github.com/Azure/azure-rest-api-specs/issues/5249
-			// "tags": tags.Schema(),
+			"identity": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Computed: true,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"principal_id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+
+						"tenant_id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+
+						"type": {
+							Type:             schema.TypeString,
+							Optional:         true,
+							DiffSuppressFunc: suppress.CaseDifference,
+							ValidateFunc: validation.StringInSlice([]string{
+								"SystemAssigned",
+							}, true),
+						},
+					},
+				},
+			},
+
+			"storage_authentication_type": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+				ValidateFunc: validation.StringInSlice([]string{
+					"ManagedIdentity",
+				}, true),
+			},
+
+			"tags": tags.Schema(),
 		},
 	}
 }
 
 func resourceArmMediaServicesAccountCreateUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).Media.ServicesClient
+	subscription := meta.(*clients.Client).Account.SubscriptionId
 	ctx, cancel := timeouts.ForCreateUpdate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
 	accountName := d.Get("name").(string)
 	location := azure.NormalizeLocation(d.Get("location").(string))
+	t := d.Get("tags").(map[string]interface{})
 	resourceGroup := d.Get("resource_group_name").(string)
+	id := parse.NewMediaServiceID(subscription, resourceGroup, accountName)
 
 	storageAccountsRaw := d.Get("storage_account").(*schema.Set).List()
 	storageAccounts, err := expandMediaServicesAccountStorageAccounts(storageAccountsRaw)
@@ -99,17 +140,22 @@ func resourceArmMediaServicesAccountCreateUpdate(d *schema.ResourceData, meta in
 			StorageAccounts: storageAccounts,
 		},
 		Location: utils.String(location),
+		Tags:     tags.Expand(t),
 	}
 
-	if _, e := client.CreateOrUpdate(ctx, resourceGroup, accountName, parameters); e != nil {
-		return fmt.Errorf("Error creating Media Service Account %q (Resource Group %q): %+v", accountName, resourceGroup, e)
+	if _, ok := d.GetOk("identity"); ok {
+		parameters.Identity = expandAzureRmMediaServiceIdentity(d)
 	}
 
-	service, err := client.Get(ctx, resourceGroup, accountName)
-	if err != nil {
-		return fmt.Errorf("Error retrieving Media Service Account %q (Resource Group %q): %+v", accountName, resourceGroup, err)
+	if v, ok := d.GetOk("storage_authentication"); ok {
+		parameters.StorageAuthentication = media.StorageAuthentication(v.(string))
 	}
-	d.SetId(*service.ID)
+
+	if _, e := client.CreateOrUpdate(ctx, id.ResourceGroup, id.Name, parameters); e != nil {
+		return fmt.Errorf("creating Media Service Account %q (Resource Group %q): %+v", id.Name, id.ResourceGroup, e)
+	}
+
+	d.SetId(id.ID(""))
 
 	return resourceArmMediaServicesAccountRead(d, meta)
 }
@@ -132,7 +178,7 @@ func resourceArmMediaServicesAccountRead(d *schema.ResourceData, meta interface{
 			return nil
 		}
 
-		return fmt.Errorf("Error retrieving Media Services Account %q (Resource Group %q): %+v", id.Name, id.ResourceGroup, err)
+		return fmt.Errorf("retrieving Media Services Account %q (Resource Group %q): %+v", id.Name, id.ResourceGroup, err)
 	}
 
 	d.Set("name", id.Name)
@@ -141,17 +187,20 @@ func resourceArmMediaServicesAccountRead(d *schema.ResourceData, meta interface{
 		d.Set("location", azure.NormalizeLocation(*location))
 	}
 
-	if props := resp.ServiceProperties; props != nil {
+	props := resp.ServiceProperties
+	if props != nil {
 		accounts := flattenMediaServicesAccountStorageAccounts(props.StorageAccounts)
 		if e := d.Set("storage_account", accounts); e != nil {
-			return fmt.Errorf("Error flattening `storage_account`: %s", e)
+			return fmt.Errorf("flattening `storage_account`: %s", e)
 		}
+		d.Set("storage_authentication_type", string(props.StorageAuthentication))
 	}
 
-	// TODO: support Tags when this bug is fixed:
-	// https://github.com/Azure/azure-rest-api-specs/issues/5249
-	// return tags.FlattenAndSet(d, resp.Tags)
-	return nil
+	if err := d.Set("identity", flattenAzureRmMediaServicedentity(resp.Identity)); err != nil {
+		return fmt.Errorf("flattening `identity`: %s", err)
+	}
+
+	return tags.FlattenAndSet(d, resp.Tags)
 }
 
 func resourceArmMediaServicesAccountDelete(d *schema.ResourceData, meta interface{}) error {
@@ -169,7 +218,7 @@ func resourceArmMediaServicesAccountDelete(d *schema.ResourceData, meta interfac
 		if response.WasNotFound(resp.Response) {
 			return nil
 		}
-		return fmt.Errorf("Error issuing AzureRM delete request for Media Services Account '%s': %+v", id.Name, err)
+		return fmt.Errorf("issuing AzureRM delete request for Media Services Account '%s': %+v", id.Name, err)
 	}
 
 	return nil
@@ -224,4 +273,30 @@ func flattenMediaServicesAccountStorageAccounts(input *[]media.StorageAccount) [
 	}
 
 	return results
+}
+
+func expandAzureRmMediaServiceIdentity(d *schema.ResourceData) *media.ServiceIdentity {
+	identities := d.Get("identity").([]interface{})
+	identity := identities[0].(map[string]interface{})
+	identityType := identity["type"].(string)
+	return &media.ServiceIdentity{
+		Type: media.ManagedIdentityType(identityType),
+	}
+}
+
+func flattenAzureRmMediaServicedentity(identity *media.ServiceIdentity) []interface{} {
+	if identity == nil {
+		return make([]interface{}, 0)
+	}
+
+	result := make(map[string]interface{})
+	result["type"] = string(identity.Type)
+	if identity.PrincipalID != nil {
+		result["principal_id"] = *identity.PrincipalID
+	}
+	if identity.TenantID != nil {
+		result["tenant_id"] = *identity.TenantID
+	}
+
+	return []interface{}{result}
 }