+++ azurerm/internal/services/postgres/postgresql_firewall_rule_resource.go
@@ -49,7 +49,7 @@ func resourceArmPostgreSQLFirewallRule() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validate.PostgreSQLServerName,
+				ValidateFunc: validate.ServerName,
 			},
 
 			"start_ip_address": {