+++ azurerm/internal/services/servicebus/data_source_servicebus_topic_authorization_rule.go
@@ -30,7 +30,7 @@ func dataSourceArmServiceBusTopicAuthorizationRule() *schema.Resource {
 			"namespace_name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validate.ServiceBusNamespaceName,
+				ValidateFunc: validate.NamespaceName,
 			},
 
 			"topic_name": {