+++ azurerm/internal/services/web/parse/slot_virtual_network_swift_connection_test.go
@@ -12,7 +12,7 @@ var _ resourceid.Formatter = SlotVirtualNetworkSwiftConnectionId{}
 
 func TestSlotVirtualNetworkSwiftConnectionIDFormatter(t *testing.T) {
 	actual := NewSlotVirtualNetworkSwiftConnectionID("12345678-1234-9876-4563-123456789012", "resGroup1", "site1", "slot1", "virtualNetwork").ID("")
-	expected := "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/networkconfig/virtualNetwork"
+	expected := "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/networkConfig/virtualNetwork"
 	if actual != expected {
 		t.Fatalf("Expected %q but got %q", expected, actual)
 	}
@@ -80,26 +80,26 @@ func TestSlotVirtualNetworkSwiftConnectionID(t *testing.T) {
 		},
 
 		{
-			// missing NetworkconfigName
+			// missing NetworkConfigName
 			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/",
 			Error: true,
 		},
 
 		{
-			// missing value for NetworkconfigName
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/networkconfig/",
+			// missing value for NetworkConfigName
+			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/networkConfig/",
 			Error: true,
 		},
 
 		{
 			// valid
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/networkconfig/virtualNetwork",
+			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/resGroup1/providers/Microsoft.Web/sites/site1/slots/slot1/networkConfig/virtualNetwork",
 			Expected: &SlotVirtualNetworkSwiftConnectionId{
 				SubscriptionId:    "12345678-1234-9876-4563-123456789012",
 				ResourceGroup:     "resGroup1",
 				SiteName:          "site1",
 				SlotName:          "slot1",
-				NetworkconfigName: "virtualNetwork",
+				NetworkConfigName: "virtualNetwork",
 			},
 		},
 
@@ -137,8 +137,8 @@ func TestSlotVirtualNetworkSwiftConnectionID(t *testing.T) {
 		if actual.SlotName != v.Expected.SlotName {
 			t.Fatalf("Expected %q but got %q for SlotName", v.Expected.SlotName, actual.SlotName)
 		}
-		if actual.NetworkconfigName != v.Expected.NetworkconfigName {
-			t.Fatalf("Expected %q but got %q for NetworkconfigName", v.Expected.NetworkconfigName, actual.NetworkconfigName)
+		if actual.NetworkConfigName != v.Expected.NetworkConfigName {
+			t.Fatalf("Expected %q but got %q for NetworkConfigName", v.Expected.NetworkConfigName, actual.NetworkConfigName)
 		}
 	}
 }